refactor(kernel): introduce kernel/ layer; relocate pure hash utility
Establishes src/kernel/ as the stdlib-only shared-utility layer (concept
architecture §2). Moves the content-hashing helper (calculateFileHash/
calculateFileHashes — only node:crypto + node:fs, genuinely pure) to
kernel/hash.ts; extract-log.ts imports it via the @/kernel alias. Test
mirror relocated to tests/unit/kernel/hash.spec.ts and repointed.

concurrency.ts was intentionally NOT moved: it depends on the logger, so
it is not stdlib-pure; relocating it would either seed a kernel→shared
dependency or require removing logging (a behavior change). Deferred.

No behavior change: build clean, 2619/2619 tests pass, lint clean.
