+++ src/kernel/hash.ts
