+++ tests/unit/stages/analyze/utils/extract-log.spec.ts
@@ -1,5 +1,6 @@
 import { readFile, stat } from 'node:fs/promises';
 
+import { calculateFileHash } from '@/kernel/hash';
 import { getCurrentSessionPaths } from '@/shared/runtime/session-context';
 import { writeMetadataFile } from '@/shared/utils/file-utils';
 import { logger } from '@/shared/utils/logger';
@@ -12,13 +13,12 @@ import {
   updateExtractLog,
   updateFileInExtractLog,
 } from '@/stages/analyze/utils/extract-log';
-import { calculateFileHash } from '@/stages/analyze/utils/hash-calculator';
 
 jest.mock('node:fs/promises');
 jest.mock('@/shared/runtime/session-context');
 jest.mock('@/shared/utils/file-utils');
 jest.mock('@/shared/utils/logger');
-jest.mock('@/stages/analyze/utils/hash-calculator');
+jest.mock('@/kernel/hash');
 
 const mockReadFile = readFile as jest.MockedFunction<typeof readFile>;
 const mockStat = stat as jest.MockedFunction<typeof stat>;
